Add CI for Float32

+++ Tests/Test Cases/MFATestCase.swift
@@ -10,7 +10,7 @@ import Foundation
 class MFATestCase {
   // Global setting for the precision used in tests.
   #if arch(arm64)
-  typealias Real = Float16
+  typealias Real = Float
   #else
   typealias Real = Float
   #endif
